requirements.txt
torch==1.2.0
torchvision==0.4.0
opencv-python==4.2.0.34
tqdm==4.32.1
matplotlib==3.1.0
numpy==1.16.4
requests==2.20.1
scikit-learn==0.21.2
Pillow==6.2.1
pyarrow==0.15.0
lz4==2.2.1
albumentations==0.4.1
pretrainedmodels==0.7.4
pandas==0.24.2
tensorboard==2.0.2
catalyst==20.3.3
ttach==0.0.2
wandb==0.9.1
ConfigArgParse==1.2.3
